+++ test.js
@@ -5,34 +5,42 @@ var rimraf = require('rimraf')
 var fixturify = require('./')
 
 
-var obj = {
-  'foo.txt': 'foo.txt contents',
-  'subdir': {
-    'bar.txt': 'bar.txt contents',
-    'symlink': ['../foo.txt']
-  }
-}
-
 test('writeSync', function (t) {
-  fs.mkdirSync('testdir1.tmp')
-  fixturify.writeSync('testdir1.tmp', obj)
-  t.deepEqual(fs.readdirSync('testdir1.tmp').sort(), ['foo.txt', 'subdir'])
-  t.deepEqual(fs.readdirSync('testdir1.tmp/subdir').sort(), ['bar.txt', 'symlink'])
-  t.equal(fs.readFileSync('testdir1.tmp/foo.txt', { encoding: 'utf8' }), 'foo.txt contents')
-  t.equal(fs.readFileSync('testdir1.tmp/subdir/bar.txt', { encoding: 'utf8' }), 'bar.txt contents')
-  t.equal(fs.readlinkSync('testdir1.tmp/subdir/symlink'), '../foo.txt')
-  rimraf.sync('testdir1.tmp')
+  rimraf.sync('testdir.tmp')
+  fs.mkdirSync('testdir.tmp')
+  fixturify.writeSync('testdir.tmp', {
+    'foo.txt': 'foo.txt contents',
+      'subdir': {
+        'bar.txt': 'bar.txt contents'
+      }
+    })
+
+  t.deepEqual(fs.readdirSync('testdir.tmp').sort(), ['foo.txt', 'subdir'])
+  t.deepEqual(fs.readdirSync('testdir.tmp/subdir').sort(), ['bar.txt'])
+  t.equal(fs.readFileSync('testdir.tmp/foo.txt', { encoding: 'utf8' }), 'foo.txt contents')
+  t.equal(fs.readFileSync('testdir.tmp/subdir/bar.txt', { encoding: 'utf8' }), 'bar.txt contents')
+
+  rimraf.sync('testdir.tmp')
   t.end()
 })
 
 test('readSync', function (t) {
-  fs.mkdirSync('testdir2.tmp')
-  // We'll be lazy and re-use the writeSync method to set up our fixture,
-  // since we tested it above
-  fixturify.writeSync('testdir2.tmp', obj)
-  var newObj = fixturify.readSync('testdir2.tmp')
-  t.deepEqual(newObj, obj)
-  rimraf.sync('testdir2.tmp')
+  rimraf.sync('testdir.tmp')
+  fs.mkdirSync('testdir.tmp')
+  fs.writeFileSync('testdir.tmp/foo.txt', 'foo.txt contents')
+  fs.mkdirSync('testdir.tmp/subdir')
+  fs.writeFileSync('testdir.tmp/subdir/bar.txt', 'bar.txt contents')
+  fs.symlinkSync('../foo.txt', 'testdir.tmp/subdir/symlink')
+
+  t.deepEqual(fixturify.readSync('testdir.tmp'), {
+    'foo.txt': 'foo.txt contents',
+      'subdir': {
+        'bar.txt': 'bar.txt contents',
+        'symlink': 'foo.txt contents'
+      }
+    })
+
+  rimraf.sync('testdir.tmp')
   t.end()
 })
 
@@ -48,6 +56,17 @@ test('error conditions', function (t) {
     t.throws(function () {
       fixturify.readSync('doesnotexist')
     })
+    t.end()
+  })
+
+  test('readSync throws on broken symlinks', function(t) {
+    rimraf.sync('testdir.tmp')
+    fs.mkdirSync('testdir.tmp')
+    fs.symlinkSync('doesnotexist', 'testdir.tmp/symlink')
+    t.throws(function() {
+      fixturify.readSync('testdir.tmp')
+    })
+    rimraf.sync('testdir.tmp')
     t.end()
   })