+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Always follow symlinks
+
 # 0.1.1
 
 * Fix bug due to missing `var`